requirements.txt
beautifulsoup4
faiss-cpu
joblib
numpy
pandas
requests
scikit-learn
sentence-transformers
trafilatura